% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparse.R
\name{argIsInit}
\alias{argIsInit}
\title{Check whether the argument parser has been initialized}
\usage{
argIsInit()
}
\value{
Logical, \code{TRUE} if \code{argParse} has been called, \code{FALSE} otherwise.
}
\description{
Check whether the argument parser has been initialized
}
