% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparse.R
\name{argPresent}
\alias{argPresent}
\title{Test whether the given option is present in the command line or not}
\usage{
argPresent(opt)
}
\arguments{
\item{opt}{Character string, option name}
}
\description{
Test whether the given option is present in the command line or not
}

