% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existArg.R
\name{existArg}
\alias{existArg}
\title{Test if named arguments exists}
\usage{
existArg(args)
}
\arguments{
\item{args}{Argument names, without leading minus sign}
}
\value{
A vector of logicals, indicating whether the arguments exist
}
\description{
Test if named arguments exists
}
\details{
Options are those arguments with a leading minus sign
(e.g. "-opt"). This function tells whether queried options exist in
the argument list.
}
\examples{
comm <- paste(c("Rscript --vanilla -e", "'", "library(ribiosArg);",
             "existArg(c(\\"opt\\", \\"opt2\\", \\"opt3\\"))", "'",
              "-opt abc -opt3"), collapse=" ")
system(comm)
}
\seealso{
\code{\link{getArg}}
}

