% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArg.R
\name{getArg}
\alias{getArg}
\title{An R-implementation of getting named aguments}
\usage{
getArg(args, onlyArg = FALSE, missingArg = FALSE)
}
\arguments{
\item{args}{Character strings, named arguments}

\item{onlyArg}{Any type, What value should be returned if only the option is available and no value has been provided}

\item{missingArg}{Any type, What value should be returned if the option is not available}
}
\value{
A list when more than one option were queried; or a vector if only one option was queried.
}
\description{
This function is out-dated. Please use \code{argparse} instead.
}
\details{
Options are those arguments with a leading minus sign. They can
have one or more values following them, which will be taken as the value
of the option. If no such values are availble, user could decide how to 
interpret the option by setting the \code{onlyArg} parameter. Similarly,
missing options can be handled by \code{missingArg}

From version 1.0.3 \code{onlyArg} and \code{missingArg} accepts \code{NULL} as inputs.
}
\seealso{
\code{\link{existArg}}
}

