% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging.R
\name{isDebugging}
\alias{isDebugging}
\title{Test whether the environment is set for debugging}
\usage{
isDebugging()
}
\value{
A logical value
}
\description{
Test whether the environment is set for debugging
}
\examples{
isDebugging()
unsetDebug()
isDebugging()
setDebug()

}
\seealso{
\code{\link{setDebug}} and \code{\link{unsetDebug}}
}
