% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging.R
\name{isIntDebugging}
\alias{isIntDebugging}
\title{Test whether the environment is set for debugging, or it's an interactive session}
\usage{
isIntDebugging()
}
\value{
A logical value
}
\description{
Test whether the environment is set for debugging, or it's an interactive session
}
\seealso{
\code{\link{isDebugging}}
}

