% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosArg.R
\name{ribiosArg}
\alias{ribiosArg}
\title{ribiosIO
ribiosIO provides Command-line argument handling for R scripting}
\description{
ribiosIO
ribiosIO provides Command-line argument handling for R scripting
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
