% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptInit.R
\name{scriptInit}
\alias{scriptInit}
\alias{initScript}
\alias{This}
\alias{function}
\alias{is}
\alias{called}
\alias{at}
\alias{the}
\alias{beginning}
\alias{of}
\alias{an}
\alias{Rscript,}
\alias{in}
\alias{order}
\alias{to}
\alias{prepare}
\alias{R}
\alias{environment}
\alias{run}
\alias{a}
\alias{script}
\alias{setting.}
\title{Prepare the environment for a script}
\usage{
scriptInit(dumpto = "ribios.dump", include.GlobalEnv = TRUE)
}
\arguments{
\item{dumpto}{Character string, file name to dump to when there is an error}

\item{include.GlobalEnv}{logical indicating if a _copy_ of the .GlobalEnv enrivonment should be included
in addition tot he 'sys.frames()'. Useful when used in a batch job}
}
\value{
Only side effect is used
}
\description{
Prepare the environment for a script
}
\examples{
\dontrun{
  scriptInit()
}
}
