% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging.R
\name{setDebug}
\alias{setDebug}
\title{Set the enrivonment for debugging}
\usage{
setDebug()
}
\value{
A logical value, whether the setting was susccessful or not
}
\description{
Set the enrivonment for debugging
}
\seealso{
\code{\link{isDebugging}} and \code{\link{unsetDebug}}
}

