% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosArg.R
\docType{package}
\name{ribiosArg-package}
\alias{ribiosArg-package}
\title{ribiosArg: Argument Handling for Command-Line, Stand-Alone R Scripts}
\description{
Provides functions to handle command-line arguments for R scripting. It enables building stand-alone R programs that accept and parse command-line options in 'BIOS' style. Zhang (2025) \url{https://github.com/bedapub/ribiosArg}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bedapub/ribiosArg}
  \item Report bugs at \url{https://github.com/bedapub/ribiosArg/issues}
}

}
\author{
\strong{Maintainer}: Jitao David Zhang \email{jitao_david.zhang@roche.com} (\href{https://orcid.org/0000-0002-3085-0909}{ORCID}) [contributor]

}
\keyword{internal}
