% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptInit.R
\name{scriptInit}
\alias{scriptInit}
\alias{initScript}
\alias{This}
\alias{function}
\alias{is}
\alias{called}
\alias{at}
\alias{the}
\alias{beginning}
\alias{of}
\alias{an}
\alias{Rscript,}
\alias{in}
\alias{order}
\alias{to}
\alias{prepare}
\alias{R}
\alias{environment}
\alias{run}
\alias{a}
\alias{script}
\alias{setting.}
\title{Prepare the environment for a script}
\usage{
scriptInit()
}
\value{
Only side effect is used
}
\description{
Prepare the environment for a script
}
\examples{
\dontrun{
  scriptInit()
}
}
