% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptSkeleton.R
\name{scriptSkeleton}
\alias{scriptSkeleton}
\title{Generate a Rscript with its skeleton}
\usage{
scriptSkeleton(file = stdout())
}
\arguments{
\item{file}{Output file. By default the function writes to standard output.}
}
\description{
Generate a Rscript with its skeleton
}
\examples{
scriptSkeleton()
}
